model.fragment: convert the value to long() in setBegin/setEnd, else it may store a float value, and it crashes in the getBegin/getEnd conversion
git-svn-id: svn+ssh://svn.gna.org/svn/advene/trunk@2481 3fb6f576-06d4-0310-ad11-a4b695aa9f05

+++ lib/advene/model/fragment.py
@@ -124,13 +124,13 @@ class AbstractNbeFragment (AbstractFragment, modeled.Modeled):
         return long(self._getModel().getAttributeNS(None, 'begin'))
     
     def setBegin(self, value):
-        return self._getModel().setAttributeNS(None, 'begin', unicode(value))
+        return self._getModel().setAttributeNS(None, 'begin', unicode(long(value)))
 
     def getEnd(self):
         return long(self._getModel().getAttributeNS(None, 'end'))
     
     def setEnd(self, value):
-        return self._getModel().setAttributeNS(None, 'end', unicode(value))
+        return self._getModel().setAttributeNS(None, 'end', unicode(long(value)))
 
     def getDuration(self):
         return self.getEnd() - self.getBegin()
